backend/requirements.txt
# Core ML Libraries
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0

# Deep Learning
torch>=2.0.0
torchvision>=0.15.0
tensorflow>=2.12.0  # or use PyTorch only

# Computer Vision
opencv-python>=4.5.0
Pillow>=9.0.0
albumentations>=1.3.0
ultralytics>=8.0.0  # YOLOv8/v10 for acne detection
git+https://github.com/THU-MIG/yolov10.git  # YOLOv10 (latest)

# Acne Detection - Roboflow Models
roboflow>=1.1.0  # For ensemble detector (acnedet-v1, skin_disease_ak, skn-1)

# Data Management
dvc[gdrive]>=3.0.0
kaggle>=1.5.0

# Experiment Tracking
wandb>=0.15.0
tensorboard>=2.12.0

# Utilities
tqdm>=4.65.0
matplotlib>=3.5.0
seaborn>=0.12.0
pyyaml>=6.0

# Development
jupyter>=1.0.0
ipython>=8.0.0
black>=23.0.0
flake8>=6.0.0
pytest>=7.0.0

# API
fastapi>=0.95.0
uvicorn>=0.21.0
streamlit>=1.20.0
python-multipart>=0.0.6
boto3>=1.26.0
google-generativeai>=0.2.0
google-cloud-vision>=3.4.0
python-dotenv>=1.0.0
